backend/requirements.txt
fastapi
uvicorn
pydantic
docker
gitpython
pyyaml
requests
